Move getTokenByAddress to tokenUtility
