+++ scripts/utils/fixEntitiesDonationCounter.js
@@ -9,26 +9,7 @@ require('../../src/models/mongoose-bn')(mongoose);
 const _groupBy = require('lodash.groupby');
 const { toBN } = require('web3-utils');
 const config = require('config');
-
-let tokensByAddress;
-
-function getTokenByAddress(address) {
-  const ANY_TOKEN = {
-    name: 'ANY_TOKEN',
-    address: '0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF',
-    foreignAddress: '0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF',
-    symbol: 'ANY_TOKEN',
-    decimals: 18,
-  };
-  if (!tokensByAddress) {
-    tokensByAddress = {};
-    config.get('tokenWhitelist').forEach(token => {
-      tokensByAddress[token.address] = token;
-    });
-  }
-  tokensByAddress[ANY_TOKEN.address] = ANY_TOKEN;
-  return tokensByAddress[address];
-}
+const {getTokenByAddress} = require('./tokenUtility')
 
 const appFactory = () => {
   const data = {};
